Move attack pipeline tests to dedicated file
